Issue #1: Move abstract classes in the parent directory.
