+++ src/Plugin/Task/ChownTask.php
@@ -4,6 +4,8 @@ declare(strict_types = 1);
 
 namespace PhpTaskman\Core\Plugin\Task;
 
+use PhpTaskman\Core\Plugin\FilesystemTask;
+
 final class ChownTask extends FilesystemTask
 {
     public const ARGUMENTS = [
