+++ src/Plugin/Task/RunTask.php
@@ -4,6 +4,7 @@ declare(strict_types = 1);
 
 namespace PhpTaskman\Core\Plugin\Task;
 
+use PhpTaskman\Core\Plugin\BaseTask;
 use PhpTaskman\Core\Robo\Task\FileEdition\LoadFileEditionTasks;
 use PhpTaskman\Core\Robo\Task\ProcessConfigFile\LoadProcessConfigFileTasks;
 use Robo\Common\BuilderAwareTrait;
@@ -16,8 +17,8 @@ final class RunTask extends BaseTask
     use LoadFileEditionTasks;
     use LoadProcessConfigFileTasks;
     use loadTasks;
-    public const ARGUMENTS = [];
 
+    public const ARGUMENTS = [];
     public const NAME = 'run';
 
     /**
