+++ src/Plugin/Task/SymlinkTask.php
@@ -4,6 +4,8 @@ declare(strict_types = 1);
 
 namespace PhpTaskman\Core\Plugin\Task;
 
+use PhpTaskman\Core\Plugin\FilesystemTask;
+
 final class SymlinkTask extends FilesystemTask
 {
     public const ARGUMENTS = [
